Move DisplayAccessStatus into its own file
