+++ common/display/src/main/scala/uk/ac/wellcome/display/models/DisplayAccessStatus.scala
@@ -1,0 +1,35 @@
+package uk.ac.wellcome.display.models
+
+import io.circe.generic.extras.JsonKey
+import io.swagger.v3.oas.annotations.media.Schema
+import uk.ac.wellcome.models.work.internal.AccessStatus
+
+@Schema(
+  name = "AccessStatus"
+)
+case class DisplayAccessStatus(
+  id: String,
+  label: String,
+  @JsonKey("type") @Schema(name = "type") ontologyType: String = "AccessStatus"
+)
+
+object DisplayAccessStatus {
+
+  def apply(accessStatus: AccessStatus): DisplayAccessStatus =
+    accessStatus match {
+      case AccessStatus.Open =>
+        DisplayAccessStatus("open", "Open")
+      case AccessStatus.OpenWithAdvisory =>
+        DisplayAccessStatus("open-with-advisory", "Open with Advisory")
+      case AccessStatus.Restricted =>
+        DisplayAccessStatus("restricted", "Restricted")
+      case AccessStatus.Closed =>
+        DisplayAccessStatus("closed", "Closed")
+      case AccessStatus.LicensedResources =>
+        DisplayAccessStatus("licensed-resources", "Licensed Resources")
+      case AccessStatus.Unavailable =>
+        DisplayAccessStatus("unavailable", "Unavailable")
+      case AccessStatus.PermissionRequired =>
+        DisplayAccessStatus("permission-required", "Permission Required")
+    }
+}
